UI: one file per chart adapter
Structure-only commit (no behavior change), per the audit-label
principle: in a repo where AI writes every edit and the owner audits
PRs by their changed-file list, file boundaries are the only
human-legible responsibility map. A diff touching charts/limits.rs
names its own blast radius; the same edit inside a 613-line charts.rs
does not.

- charts.rs (206 lines) keeps the shared column frame, the facade
  re-exports (callers keep addressing charts::*), and the frame
  contract tests.
- charts/{model_daily,hourly,limits,credits}.rs each hold exactly one
  panel. Adapter entry points are pub(in crate::ui) behind pub(super)
  re-exports, so visibility stays closed to the ui module.
- limits keeps its fixed 0-100% scale rather than adopting level_for —
  that would be a behavior change and this commit moves code only.
